Rename App and Android package name
