+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.diary",
+    package = "com.saheeranas.diaryapp",
 )
 
 android_resource(
     name = "res",
-    package = "com.diary",
+    package = "com.saheeranas.diaryapp",
     res = "src/main/res",
 )
