+++ android/app/src/main/java/com/saheeranas/diaryapp/MainActivity.java
@@ -1,4 +1,4 @@
-package com.diary;
+package com.saheeranas.diaryapp;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
    */
   @Override
   protected String getMainComponentName() {
-    return "Diary";
+    return "PrivateDiaryApp";
   }
 }
