+++ android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Diary</string>
+    <string name="app_name">Private Diary App</string>
 </resources>
