+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "Diary",
-  "displayName": "Diary"
+  "name": "PrivateDiaryApp",
+  "displayName": "Private PrivateDiaryApp App"
 }
