+++ ios/Podfile
@@ -3,7 +3,7 @@ require_relative '../node_modules/@react-native-community/cli-platform-ios/nativ
 
 platform :ios, '11.0' 
 
-target 'Diary' do
+target 'PrivateDiaryApp' do
   config = use_native_modules!
 
   use_react_native!(
@@ -12,7 +12,7 @@ target 'Diary' do
     :hermes_enabled => false
   )
 
-  target 'DiaryTests' do
+  target 'PrivateDiaryAppTests' do
     inherit! :complete
     # Pods for testing
   end
