+++ ios/PrivateDiaryApp/AppDelegate.h
