+++ ios/PrivateDiaryApp/Images.xcassets/AppIcon.appiconset/Contents.json
