+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "Diary",
+  "name": "PrivateDiaryApp",
   "version": "0.0.1",
   "private": true,
   "scripts": {
